Added search bar on home page and recent photos
Changed backend handling of get requests for photos to include filter.

Added a search page with no functionality still.

Added a recent uploads section below it.

Made custom hooks for fetching multiple photos depending on a condition.

Fixed z-index on auth modal and dropdown menu to not overlap with other components.

+++ client/src/pages/Gallery/PhotoDetails.jsx
@@ -7,16 +7,22 @@ import dateFormat from "../../utils/dateFormat";
 import useFetchUserById from "../../hooks/useFetchUserById";
 import { useUser } from "../../contexts/userContext";
 import axios from "axios";
+import DeletePhotoModal from "./DeletePhotoModal";
+
 
 
 function PhotoDetail() {
   const [isOwner, setIsOwner] = useState(false);
+  const [isModalOpen, setIsModalOpen] = useState(false);
   const { photoId } = useParams();
   const { photo, isLoading } = useFetchSingle(photoId);
   const { user } = useFetchUserById(photo?.user);
 
   const { user: loggedInUser, token } = useUser();
   const navigate = useNavigate();
+
+ 
+  
 
 
   useEffect(() => {
@@ -31,26 +37,20 @@ function PhotoDetail() {
 
 
   const handleDelete = async (photoId) => {
+    if (!photoId) return;
+  
     try {
       await axios.delete(`http://localhost:3000/api/photos/${photoId}/delete`, {
         headers: {
-            Authorization: `Bearer ${token}`, // Attach token
+          Authorization: `Bearer ${token}`,
         },
-    })
-
-        // Update the UI by removing the deleted photo
-        setPhotos((prevPhotos) => prevPhotos.filter(photo => photo._id !== photoId));
-
-        // Update the photo count
-        setPhotosCount((prevCount) => prevCount - 1);
-
-        navigate('/gallery')
-
+      });
+  
+      navigate('/user-gallery');
     } catch (error) {
-        console.error(`Cannot delete photo: ${error.message}`);
+      console.error(`Cannot delete photo: ${error.message}`);
     }
-};
-
+  };
 
 
   if (isLoading) return <p>Loading photo...</p>;
@@ -77,12 +77,20 @@ function PhotoDetail() {
 
                   {/* X Button in the top-right corner */}
                   {isOwner && (
+                    <>
                     <button
-                    onClick={() => handleDelete(photo._id)} // Add your delete functionality
+                    onClick={() => setIsModalOpen(true)} 
                     className="text-gray-500 hover:text-gray-800"
                   >
                     X
                   </button>
+
+                  <DeletePhotoModal
+                    isOpen={isModalOpen}
+                    onClose={() => setIsModalOpen(false)}
+                    onDelete={() => handleDelete(photoId)}
+                  />
+                  </>
                   )}
                  
                 </h1>
@@ -125,8 +133,12 @@ function PhotoDetail() {
               </aside>
             </div>
 
+          
+              <CommentsSection photo={photo} />
+
+
             {/* Bottom Section: Comments */}
-            <CommentsSection photo={photo} />
+            
             {/* Comment listing here */}
           </div>
         </div>
